Use 64 bit for SSD computation

+++ src/measures/ssd.py
@@ -12,8 +12,8 @@ def compute_ssd(target, warped, mask):
     :param mask: The image mask.
     :return: The computed sum of square difference
     """
-    warped_data = warped.data.astype(np.float32)
-    target_data = target.data.astype(np.float32)
+    warped_data = warped.data.astype(np.float64)
+    target_data = target.data.astype(np.float64)
     valid_voxels = np.count_nonzero(mask.data)
     ssd = np.sum(((warped_data - target_data)**2))/valid_voxels
     return ssd
